services/n8n_obsidian_api/requirements.txt
fastapi~=0.109.0
httpx~=0.26.0
opentelemetry-api~=1.22.0
opentelemetry-exporter-otlp~=1.22.0
opentelemetry-instrumentation-fastapi~=0.43b0
opentelemetry-instrumentation-httpx~=0.43b0
opentelemetry-sdk>=1.22.0
pydantic-settings~=2.1.0
pydantic~=2.5.0
pyyaml~=6.0.1
uvicorn[standard]~=0.27.0